refactor(traces): shared TraceListPanel + TraceDetailCard, reused in Trace inspect
Lift the per-layer Traces tab's list (rail) and rich detail card (sw-card:
trace-id header + id/url copy + Default/Tree/Statistics view modes + KPIs
+ service legend + span-detail modal) into render/widgets/, consumed by
BOTH the Traces tab and the cross-layer Trace inspect view — so they
render identically by construction. LayerTracesView shrinks ~1700 lines
(pure move); the Esc cascade now routes through the card's modal-open
report. No BFF/wire changes. type-check + lint + build green.
